Create ARDUINO LPG PROJECT CODE
Today, Liquefied Petroleum Gas ( LPG) is the most common cooking fuel throughout our country. With the increase in the number of people using this fuel, it is the need of the hour to provide some safety standards.
Gas leakage is a major problem in the industrial sector, residential premises, etc. Here is gas leakage detector which is used to detect the leakage of gas and it can protect us from gas damage.
One of the preventive methods to stop accidents associated with the gas leakage is to install a gas leakage detection kit at vulnerable places.
The aim of this paper is to propose and discuss a design of a gas leakage detection system that can automatically detect, alert and control gas leakage.